Move ParsableVariable to ncore
