+++ key.core/src/main/java/de/uka/ilkd/key/logic/op/LogicVariable.java
@@ -8,6 +8,7 @@ import java.util.Objects;
 import de.uka.ilkd.key.ldt.JavaDLTheory;
 
 import org.key_project.logic.Name;
+import org.key_project.logic.ParsableVariable;
 import org.key_project.logic.sort.Sort;
 import org.key_project.util.EqualsModProofIrrelevancy;
