+++ key.core/src/main/java/de/uka/ilkd/key/logic/op/QuantifiableVariable.java
@@ -3,6 +3,7 @@
  * SPDX-License-Identifier: GPL-2.0-only */
 package de.uka.ilkd.key.logic.op;
 
+import org.key_project.logic.ParsableVariable;
 import org.key_project.logic.op.SortedOperator;
 import org.key_project.util.EqualsModProofIrrelevancy;
