+++ key.core/src/main/java/de/uka/ilkd/key/logic/op/SchemaVariable.java
@@ -7,6 +7,7 @@ package de.uka.ilkd.key.logic.op;
 import de.uka.ilkd.key.util.pp.Layouter;
 
 import org.key_project.logic.Named;
+import org.key_project.logic.ParsableVariable;
 
 /**
  * This interface represents the root of a schema variable hierarchy to be express termstructures
