+++ key.core/src/main/java/de/uka/ilkd/key/logic/op/UpdateableOperator.java
@@ -3,6 +3,8 @@
  * SPDX-License-Identifier: GPL-2.0-only */
 package de.uka.ilkd.key.logic.op;
 
+import org.key_project.logic.ParsableVariable;
+
 /**
  * Operators implementing this interface may stand for locations as well. This means e.g. occur as
  * top level operators on the left side of an assignment pair of an update.
