+++ key.core/src/main/java/de/uka/ilkd/key/nparser/varexp/ArgumentType.java
@@ -5,9 +5,9 @@ package de.uka.ilkd.key.nparser.varexp;
 
 import de.uka.ilkd.key.java.abstraction.KeYJavaType;
 import de.uka.ilkd.key.logic.Term;
-import de.uka.ilkd.key.logic.op.ParsableVariable;
 import de.uka.ilkd.key.rule.conditions.TypeResolver;
 
+import org.key_project.logic.ParsableVariable;
 import org.key_project.logic.sort.Sort;
 
 /**
