+++ key.core/src/main/java/de/uka/ilkd/key/smt/newsmt2/SeqDefHandler.java
@@ -21,6 +21,7 @@ import de.uka.ilkd.key.logic.op.*;
 import de.uka.ilkd.key.smt.SMTTranslationException;
 import de.uka.ilkd.key.smt.newsmt2.SExpr.Type;
 
+import org.key_project.logic.ParsableVariable;
 import org.key_project.logic.sort.Sort;
 import org.key_project.util.collection.DefaultImmutableSet;
 import org.key_project.util.collection.ImmutableSet;
