+++ key.ncore/src/main/java/org/key_project/logic/ParsableVariable.java
@@ -1,7 +1,7 @@
 /* This file is part of KeY - https://key-project.org
  * KeY is licensed under the GNU General Public License Version 2
  * SPDX-License-Identifier: GPL-2.0-only */
-package de.uka.ilkd.key.logic.op;
+package org.key_project.logic;
 
 /**
  * This interface represents the variables that can be recognized by one of the parsers.
